move namedValueToValue to utils_go18.go
